Add foundation of logger and error-interceptor
This commit adds the foundation of `@woodwork/logger` and
`@woodwork-interceptor-error`.

`@woodwork/logger` being the primary idea to allow for sending frontend
logs to a backend service to consume them.
The `@woodwork/interceptor-error` is what made me want to make a package
out of both, I want `@woodwork/logger` to stay minimal and allow for
"plugins" to be added, depending on how you send your objects
non-enumerable properties wouldn't be added to the request. This is why
I decided to create `@woodwork/interceptor-error` to extract necessary
properties, to better know where and what went wrong on the frontend but
also enable for consumption of relevant data. How often did this happen?
Did a certain error increase after a deploy, what might be the cause of
it? Etc..

@woodwork/logger @woodwork/interceptor-error